requirements.txt
fixedint
libusb1
matplotlib
opencv-python
websockets
